requirements.txt
streamlit==1.45.1
fastapi==0.115.12
uvicorn==0.34.2
requests==2.32.3
python-dotenv==1.1.0
weaviate-client==4.6.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-weaviate==0.0.4
requests==2.32.3
beautifulsoup4==4.12.3
html5lib==1.1
sentence-transformers==3.0.1
torch>=2.2.0